fix the no results message to return as json
